+++ examples/middleware/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @stackframe/example-middleware-demo
 
+## 2.8.52
+
+### Patch Changes
+
+- @stackframe/stack@2.8.52
+
+## 2.8.51
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack@2.8.51
+
+## 2.8.50
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack@2.8.50
+
+## 2.8.49
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack@2.8.49
+
 ## 2.8.48
 
 ### Patch Changes