+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @stackframe/stack-ui
 
+## 2.8.52
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.52
+
+## 2.8.51
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.51
+
+## 2.8.50
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.50
+
+## 2.8.49
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.49
+
 ## 2.8.48
 
 ### Patch Changes