+++ packages/template/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @stackframe/stack
 
+## 2.8.52
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.52
+  - @stackframe/stack-ui@2.8.52
+  - @stackframe/stack-sc@2.8.52
+
+## 2.8.51
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.51
+  - @stackframe/stack-ui@2.8.51
+  - @stackframe/stack-sc@2.8.51
+
+## 2.8.50
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-sc@2.8.50
+  - @stackframe/stack-shared@2.8.50
+  - @stackframe/stack-ui@2.8.50
+
+## 2.8.49
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-sc@2.8.49
+  - @stackframe/stack-shared@2.8.49
+  - @stackframe/stack-ui@2.8.49
+
 ## 2.8.48
 
 ### Patch Changes